requirements.txt
openai-whisper
jupyterlab
ipywidgets
transformers
sentencepiece
widgetsnbextension
tqdm
moviepy
pyyaml